+++ src/YAYL/YamlPathType.cs
@@ -1,0 +1,7 @@
+namespace YAYL;
+
+public enum YamlFilePathType
+{
+    RelativeToFile,
+    RelativeToCurrentDirectory,
+}